Actualización completa del backend CafeAppBackend

+++ CafeAppBackend/src/main/java/com/cafeapp/backend/repositorio/CarritoRepository.java
@@ -6,6 +6,18 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CarritoRepository extends JpaRepository<Carrito, Integer> {
+/**
+ * Repositorio para la entidad {@link Carrito}.
+ *
+ * Permite obtener el carrito asociado a un usuario.
+ */
+public interface CarritoRepository extends JpaRepository<Carrito, Long> {
+
+    /**
+     * Busca el carrito asociado a un usuario.
+     *
+     * @param usuario Usuario propietario del carrito
+     * @return Optional con el carrito si existe
+     */
     Optional<Carrito> findByUsuario(Usuario usuario);
 }
