+++ CafeAppBackend/src/main/java/com/cafeapp/backend/repositorio/CentroRepository.java
@@ -3,7 +3,8 @@ package com.cafeapp.backend.repositorio;
 import com.cafeapp.backend.modelo.Centro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
-public interface CentroRepository extends JpaRepository<Centro, Integer> {}
-
+/**
+ * Repositorio para la entidad {@link Centro}.
+ */
+public interface CentroRepository extends JpaRepository<Centro, Long> {
+}
