+++ CafeAppBackend/src/main/java/com/cafeapp/backend/repositorio/CursoRepository.java
@@ -3,6 +3,8 @@ package com.cafeapp.backend.repositorio;
 import com.cafeapp.backend.modelo.Curso;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CursoRepository extends JpaRepository<Curso, Integer> {}
-
-
+/**
+ * Repositorio para la entidad {@link Curso}.
+ */
+public interface CursoRepository extends JpaRepository<Curso, Long> {
+}
