+++ CafeAppBackend/src/main/java/com/cafeapp/backend/repositorio/ItemCarritoRepository.java
@@ -3,5 +3,8 @@ package com.cafeapp.backend.repositorio;
 import com.cafeapp.backend.modelo.ItemCarrito;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemCarritoRepository extends JpaRepository<ItemCarrito, Integer> {}
-
+/**
+ * Repositorio para la entidad {@link ItemCarrito}.
+ */
+public interface ItemCarritoRepository extends JpaRepository<ItemCarrito, Long> {
+}
