+++ CafeAppBackend/src/main/java/com/cafeapp/backend/repositorio/RolRepository.java
@@ -1,10 +1,10 @@
 package com.cafeapp.backend.repositorio;
 
-
 import com.cafeapp.backend.modelo.Rol;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RolRepository extends JpaRepository<Rol, Integer> {}
-
-
-
+/**
+ * Repositorio para la entidad {@link Rol}.
+ */
+public interface RolRepository extends JpaRepository<Rol, Long> {
+}
