+++ CafeAppBackend/src/main/java/com/cafeapp/backend/repositorio/TurnoRepository.java
@@ -3,5 +3,8 @@ package com.cafeapp.backend.repositorio;
 import com.cafeapp.backend.modelo.Turno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TurnoRepository extends JpaRepository<Turno, Integer> {
+/**
+ * Repositorio para la entidad {@link Turno}.
+ */
+public interface TurnoRepository extends JpaRepository<Turno, Long> {
 }
